Add carnot test for empty tables.
Summary: This test makes sure that Carnot will not exit if it has an empty table. This is a test I wrote to debug the Spinnaker issue. It didn't turn out to be the root cause, but might as well throw it in to prevent a regression that makes empty tables problematic.

Test Plan: this is a test!!

Reviewers: nserrino, michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3156

GitOrigin-RevId: f3cbe0c8a7837a843796a1e81a286949d0489739