Update Agent Metadata operations to use new cache kvstore
Summary:
replacing the EtcdMetadataStore with our new Cached kvstore is an incremental process.
There are two places we currently use EtcdMetadataStore:
- agent.go
- metadata_handler.go
But, there are actually parts of agent.go that directly interface with the etcd client itself... which is bad.
So, the first step of this process is to move the agent creation/deletion/updating functionality to the cached kvstore.

This diff only implements the new metadata store for the agent operations. it does not yet replace anything in the existing codebase yet.

Test Plan:
unit test
,

Reviewers: zasgar, nserrino, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3115

GitOrigin-RevId: 2d9a92af6881f23627a42a50ab4c5e585c61698d