[CLEANUP] Replace disabled_ with disabled().
Summary:
I am toying with the idea of having a generic State:
kUnknown, kTransferring, kDisabled
to indicate the state of the ConnectionTracker.

The previous state_ would be changed to parsing_state_ as it is related
to the state of parsing the content of the events.

Anyhow, this diff is trivial, and would be useful regardless.

Test Plan: Jenkins

Reviewers: oazizi, #engineering

Reviewed By: oazizi, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3153

GitOrigin-RevId: cc756640c2671d66aebea49117eebd1d5ea759a1