Make ComputedSchema a single message to store with a single key in etcd
Summary:
If we had a long running etcd database and something happened where part of the schema disappeared from the agents, the missing tables wouldn't be removed and caused
crashes that were rather cryptic. This fixes that problem by combining Schema into one message and updating the entire schema with a single write - no need to delete anything.

This is really a temporary solution until we figure out a better method of knowing which data is available on which Agents. This will be addressed in our discussion of how Kelvin backup nodes will work.

Test Plan: Updated etcd tests to make sure the new storage method works, tested on Spinnaker cluster, update misc tests.

Reviewers: michelle, zasgar, nserrino, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3155

GitOrigin-RevId: 809aafca34ca2d6bd3152b39db1ac0d54df8e34e