Change ResolveExe() to return exe symlink if the target PIDs are not inside container
Summary:
This applies to the cases where:
- Stirling and target processes runs inside the same container, eg., tests.

This is a workaround to PL-1249. I spent some time, and could not figure out the root cause.
But the test coverage on gRPC uprobing is more important. So I opt to add this.

Follow up is to update the grpc_trace_bpf_test.cc to test the actual content of the uprobe data.

Test Plan: Jenkins

Reviewers: oazizi, zasgar, #engineering

Reviewed By: oazizi, #engineering

JIRA Issues: PL-1249

Differential Revision: https://phab.corp.pixielabs.ai/D3058

GitOrigin-RevId: c764f8eb69b0377354a208d2ecc358410333fcbe