[CLEANUP] Refactor APIs to be used inside SocketTraceConnector.
Summary:
A draft of the eventual changes for checking remote IP address of a
record and the cluster CIDR is:
https://phab.corp.pixielabs.ai/differential/diff/11447

This change is created such that the interface works around a new
struct IPAddress, which is similar to CIDRBlock, and unifies IPv4 and
IPv6 addresses.

The old IPv4Address and IPv6Address are likely not going to be used.
But they are kept for later cleanup, as the details are not yet fully
revealed.

Test Plan: Jenkins

Reviewers: oazizi, #engineering

Reviewed By: oazizi, #engineering

JIRA Issues: PL-1291

Differential Revision: https://phab.corp.pixielabs.ai/D3143

GitOrigin-RevId: 378af438bb2d86ab8246634443ee9fbb882026ee