Socket Prober: Proof of concept with network namespaces
Summary:
This is part 1 of supporting network namespaces.

A test shows that it is possible to capture connections of other network namespaces (as long as you have root access).

Next step is to provide a more friendly interface to automatically switch namespaces.

Test Plan:
Main part is a new test.

Also, new tests added for supporting changes to the baseline code.

Reviewers: zasgar, yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3148

GitOrigin-RevId: 6c67a0622f568564e00a72159dd8ab030233bc12