PL-1319 part 1: refactor ResolveColumn into ResolveColumnType and ResolveColumnIndex
Summary: This refactor is in service of PL-1319, which will add an analyzer phase that prunes unused columns from the plan. In order to do that, we want to defer setting the column index as long as possible, because the input/output relation of each operator may change when its columns are pruned. During the analyzer phase, we want to entirely deal in terms of column names, rather than mixing both, only moving over to column indexes at the end. Right now, column type and column index were being set at the same point, but column type is needed for many phases of the analyzer, and column index is only needed at the very end when ToProto is called on the operators. Therefore, these get set separately and I added a rule where the column index is resolved at the final step of the analyzer.

Test Plan: existing should pass

Reviewers: philkuz, zasgar, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PL-1319

Differential Revision: https://phab.corp.pixielabs.ai/D3171

GitOrigin-RevId: 11bcb7fe9f663a91097a2c153e78fef555d53eac