Network namespace scanner
Summary:
This small utility scans for all network namespaces.

It creates a map of network inode number to all the pids that belong to that namespace.

I am including all the PIDs, because PIDs can be unstable and disappear, so the caller may want some back-up PIDs to identify the network namespace by.

As a future direction, may modify this so that it operates on the network namespace directly, so that state doesn't need to be stored in a map.

Test Plan: Test added.

Reviewers: yzhao, zasgar, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PL-1333

Differential Revision: https://phab.corp.pixielabs.ai/D3181

GitOrigin-RevId: a2e66e1f02481f49ea918a06be11e83a17fe0a8e