[PL-1323] Update VarTable data structure to a heirarchical structure
Summary: This change brings about support for functions and makes sure that we can lookup values in the hierarchy. Var tables can make children that contain their own local state. When we lookup variables, we now check the local state first, then the parent's state and so on until we either find the object or reach the top most table in the hierarchy.

Test Plan: n/a, next diff has a test of this.

Reviewers: zasgar, michelle, nserrino, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3196

GitOrigin-RevId: c683ca3138c7cb6d9ac109f73540d67e40eeb882