PL-1319: Refactor Rule into Rule and DistributedRule
Summary:
PL-1319 as well as PL-1317 will require the addition of a distributed analyzer, which will share many characteristics with the current analyzer.  We want to be able to have the same logic that exists for rules on IRs be able to execute on DistributedPlans as well.

Analyzer:
traverses IR, applies Rules to IRNodes

DistributedAnalyzer:
traverses DistributedPlan, applies DistributedRules to CarnotInstances.

We want the rule executor and the graph walking stuff, as well as the patterns for Apply() on Rules to be shared on both these cases.
The first use case for a DistributedAnalyzer is to have IR rules executed on each of the Plans for each CarnotInstance, so support for
that kind of rule is added in this diff.
However there will be other kinds of DistributedRules that actually modify the top-level DAG of DistributedPlan in the future. For example,
an optimization that removes an entire CarnotInstance from the DistributedPlan because a filter condition causes there to be no data present
on a given node that will be included on the output.

Next step will be to edit RuleExecutor to be generalized to be able to use Rules or DistributedRules.
After that, the DistributedAnalyzer will be added to the codebase.
Then, a rule which needs to be added for unions post-distributed splitting will be added to the DistributedAnalyzer.
Eventually, much of the distributed splitting/stiching logic can be re-articulated as part of the DistributedAnalyzer.

Test Plan: existing, added

Reviewers: philkuz, zasgar, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PL-1317, PL-1319

Differential Revision: https://phab.corp.pixielabs.ai/D3191

GitOrigin-RevId: 6e42870fbb51c0454a6248c078c1011b71ae05a9