PL-1319: Templatize the rule executor in preparation for the DistributedAnalyzer
Summary: Depends on D3191. We want to share logic in the RuleExecutor that is currently tied to the IR graph only, with the future DistributedAnalyzer which will operate on the DistributedPlan in a similar way. As a result, we follow the pattern of D3191 and make RuleExecutor templatized on the input Graph type. This way, the analyzer can inherit from RuleExecutor<IR> and the DistributedAnalyzer can inherit from RuleExecutor<DistributedPlan>.

Test Plan: existing since this is a refactor

Reviewers: philkuz, zasgar, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PL-1317, PL-1319

Differential Revision: https://phab.corp.pixielabs.ai/D3193

GitOrigin-RevId: d145031b353bb637f65c3f5c28e3c81d1db35362