Add compiler support for UDTF defaults
Summary: UDTF defaults can now be specified and passed into the compiler. This approach is a bit hacky because we have to convert it into an expression that the ast visitor interprets instead of raw values. The main place this becomes an issue is the uint  because it converts the uint  into a string which will then be converted back into a uint. There's probably a solution to fix this by passing in IRNodes isntead of strings into the default. Will investigate in PL-1335

Test Plan: Added test to make sure that default values show up.

Reviewers: zasgar, michelle, nserrino, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3208

GitOrigin-RevId: 80ee503aea82c152cbbd4e659dfd0a3d08642410