Fix agent schemas
Summary:
We recently updated how computedSchemas works to fix a bug that occurs when some agents don't share the same schema.
this resulted in a bug where we couldn't run any queries because the compiler couldnt find any tables.
it turns out that Kelvin (which somehow always connects to the metadata service last), was sending empty schema udpates to the metadata service and the metadata service would update its computedSchemas to be empty.

Test Plan: ran skaffold

Reviewers: zasgar, philkuz, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3201

GitOrigin-RevId: b20a5ab373c9dbc940736a46cf27e365296c4810