PL-1319: Refactor UnionIRs in preparation for column pruning rule
Summary:
PL-1319 will add in a compiler optimization that prunes columns that are unnecessary to the script output.
In order to do this nicely, it makes sense for the resolution of a column to its ultimate index in the plan to happen as late in the game as possible, as the input relation to a given operator may shift as columns are pruned.
UnionIR previously stored column indexes to refer to columns, rather than a ColumnIR.
This is error prone given the upcoming optimization, because those indexes could become stale. We want all IRNodes to use ColumnIR types when referring to columns so that it's easier for us to figure out what is able to be pruned.
As a result, I moved UnionIR to using ColumnIR instead of indexes.
This exposed an issue where we were not resolving column indexes for columns/operators added after the analyzer phase.
As a result, a step in the distributed analyzer was added to resolve column indexes, so that nodes that were added to the plan after the initial analyzer phase still have their columns resolved.

Depends on D3171 and D1319.

Test Plan: added tests

Reviewers: philkuz, zasgar, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PL-1317, PL-1319

Differential Revision: https://phab.corp.pixielabs.ai/D3202

GitOrigin-RevId: 4b6c3ef40d3a44b1bcd32d0500682d5ff2f6fdf5