PL-1319: Add another helper method RequiredInputColumns to OperatorIR.
Summary: PL-1319 will add a rule that prunes unneeded columns from the plan in the analyzer phase. this change provides a helper function for that rule which lives on the IR node. Each IR node can use RequiredInputColumns to traverse its internals and return the set of columns (for each input parent) that it needs to materialize its data.

Test Plan: added

Reviewers: philkuz, zasgar, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PL-1319

Differential Revision: https://phab.corp.pixielabs.ai/D3230

GitOrigin-RevId: 4d39e4320a3400e0347f641c4fcbcc5331fd83c7