SocketProberManager
Summary:
This diff adds a SocketProberManager, and integrates into Stirling.

With this diff, Stirling is now able to resolve endpoints that are not in its own network namespace.

Test Plan:
Testing with network namespaces is a pain.

Still, added a test for SocketProberManager. Only caveat is that must be run manually (doesn't work on Jenkins).

Also tested on GKE, and appears to work.

Evaluating options for testing the Stirling integration.

Reviewers: zasgar, yzhao, #engineering

Reviewed By: yzhao, #engineering

JIRA Issues: PL-1333

Differential Revision: https://phab.corp.pixielabs.ai/D3212

GitOrigin-RevId: 56f8524f3be4cf48d953a452493848f3cef4bda1