Update agents to use new metadata store
Summary:
Now that the metadata store refactor has all of the old functionality, we can start to update our code to actually use the new metadatastore.
this diff updates the agents code to use the new metadata store. since the interface for the metadata store has changed slightly, we will keep the old interface around for now, since that interface is still being used in metadata_handler. it will be deleted in the next diff.

Test Plan: ran skaffold. everything appears to be heartbeating as normal, and queries still seem to work as normal.

Reviewers: zasgar, nserrino, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3203

GitOrigin-RevId: 990812e4ad5ee462156f5c641ea94e720cb519fe