PL-1319: Add helper method for column pruning rule to ir_nodes
Summary:
With PL-1319, there will be a rule that removes unused columns from the
output plan. It will specify to IR nodes which columns they need to keep
based on which outputs of theirs other nodes are using.

This change provides a function to IR nodes, PruneOutputColumnsTo()
which removes columns from an IR node that are not needed to produce the
specified columns that are passed into the function. As a result, if a
node only needs to output "bar", it currently produces "abc", "foo", and
"bar", but it needs to output "foo" in order to output "bar", the
columns would be pruned to "foo" and "bar". (This "foo" case only really comes up
for aggregates, where removing group by columns affects the output
result.)

It is provided as a member function of IRNode because it sometimes involves
modifying internal state of the node, and also to ensure that every new
node will not forget to implement this functionality.

Test Plan: added tests

Reviewers: philkuz, zasgar, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PL-1319

Differential Revision: https://phab.corp.pixielabs.ai/D3224

GitOrigin-RevId: 57f4ff4e522b1344eb22e4598cb810469f5d133c