Fix expired session bug
Summary:
users were not able to log in when their previous session has expired.
that is because Gorilla throws an error when it gets the default session, and we just throw an error after seeing Gorilla has thrown an error.
however, Gorilla's cookieStore.Get() will always return a valid session even if it has thrown an error. in the login case, we don't mind if the previous session was invalid/expired, because we will overwrite the session with new values anyway.

Test Plan: n/a

Reviewers: zasgar, malthus, #engineering

Reviewed By: malthus, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3253

GitOrigin-RevId: 42b5455e7e7c31a11e6b57f2e70edc230472ec57