PL-1330: Collections (Lists & Tuples) should not be Expressions
Summary:
As we add in Unions to the QL (df.append), it will take a list of dataframes.
df.append([foo,bar]) or df.append(foo). For the former case, we want to use the ListIR node which breaks the
requirement that Lists/Collections are expressions, because a list of operators cannot be an expression.
Since lists don't actually need to be expressions in our codebase, let's make collections its own category.

Test Plan: existing since this is a refactor

Reviewers: philkuz, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PL-1330

Differential Revision: https://phab.corp.pixielabs.ai/D3259

GitOrigin-RevId: bb2ae4c6ecc265107dd39d7b87240c95b0fd0978