PL-1349: Add rule to clean up stray IRNodes that are not connected to an operator.
Summary:
We don't want to have random stray nodes that are lying around and not used by the plan.
We can consider adding it as a DCHECK to other rule batches or the ast_visitor code if we don't want to rely on it for pruning and instead expect the other rules to take care of themselves. In that case we would run the rule and DCHECK that it didn't do any work on the graph, and if it did we would know that we were leaving stray nodes.

Test Plan: added

Reviewers: philkuz, zasgar, jamesbartlett, #engineering

Reviewed By: philkuz, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3271

GitOrigin-RevId: 83f0db4325f0aa6228e9190af9de86c526ec814f