Stirling bug-fix: Restore HTTP2 data pushes
Summary:
The change that introduced the trackers in kCollecting state introduced a bug where data was not being pushed to the tables.

This is because Uprobes were not setting their roles in the same way as other protocols (since it is hard to infer in BPF).

This diff moves the HTTP2 role inference code so that kicks in before the push into the tables, thus unblocking the HTTP2 data.

lint

Test Plan:
We really need an end-to-end test for things like this. Will consider expanding stirling_wrapper tests as future work.

The other entity that could have caught this is grpc_trace_bpf_test, but the uprobe tests are disabled. Will look into that too, when time permits.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

JIRA Issues: PL-1364

Differential Revision: https://phab.corp.pixielabs.ai/D3270

GitOrigin-RevId: 2753cd26d59557cc5feb2b0116dd5a11d353a496