[CLEANUP] Remove bcc_bpf_interface/http2.h
Summary:
This was created to support tracking event's offset in the connection's data stream,
and to aid recovering from lost events.

This was never implemented though.

And there is a more generic approach, which is to record data event offset directly, and correctly
identify gaps caused by lost events in user space.

Test Plan: Jenkins

Reviewers: oazizi, #engineering

Reviewed By: oazizi, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3286

GitOrigin-RevId: 335fc94fef410fb6a6901ba9085b9d0fa9b9080a