Misc compiler cleanup
Summary: DeleteNodeAndChildren is not used and it is also not safe to nodes with multiple parents (ExpressionIR with multiple parents got introduced after it was written). Also DeleteNode was leaving the nodes in id_to_nodes_map which led to inconsistent results between the IR dag and IR node map (each of which can be checked depending on the situation).

Test Plan: edited a test to work with these fixes

Reviewers: philkuz, jamesbartlett, #engineering

Reviewed By: philkuz, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3283

GitOrigin-RevId: 24290054ab30dc4454a300562a65eae65428d6b2