Cleanup: Make DataFrame an attribute of Pixie instead of a method.
Summary: This cleans up DataFrame to make type checking easier. Moved the Dataframe method from pixie to the constructor of Dataframe(). It was clear we had some unused code in the pixie module logic w.r.t GetAttributeImpl. Removing the original body and updating the attributes made it less prone to potential bugs.

Test Plan: added a test for the new dataframe constructor

Reviewers: nserrino, jamesbartlett, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3297

GitOrigin-RevId: 924c74e4a373d1a6ac4bf355bf562cc60b5a7c74