MDS crash: fix bug where etcd gets out of sync when mds restarts
Summary:
the latest MDS refactor was causing constant PEM/kelvin crashes when the metadata service restarts. this is especially prevalent when running vizier on skaffold.
after looking through logs Omid provided, this is what happened:
- metadata pod A is running
- metadata pod A is killed (but takes some time for it to actually terminate)
- new metadata pod B is started
- metadata pod B tries to become leader, but metadata pod A has not relinquished leadership because it is still dying
- metadata pod B receives updates, but ignores them because it is not the leader. most notably, it receives an update for the Metadata Service saying it is now served by pod B. metadata in etcd is now out of sync, because it still thinks metadata service is served by pod A.
- metadata pod B finally becomes leader
- metadata pod A finally dies, metadata pod B receives the pod update and writes it to etcd
- agent starts up, metadata B sends it updates, including information about the metadata service which still references pod A. however, it does not actually send information about pod A, since it is dead.
- agent crashes because it doesn't know anything about pod A.

To handle recovery, the non-leader metadata instances should not ignore k8s updates. now that we have the cache, we can keep the updates in memory. so, if a metadata service becomes leader it can flush any messages that were missed by the pod that was once the leader. if the pod never becomes leader, we should periodically clear out the cache (without flushing to etcd), because we can assume that the data has already made it safely to etcd by the leader.
this diff only handles the problem where etcd gets out of sync during a metadata crash.
in a follow up diff, we still need to address the problem where the updates we send to the agent will get out of sync.

Test Plan: ran skaffold, updated metadata service code, made sure that the correct updates are flushed to etcd.

Reviewers: zasgar, oazizi, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3280

GitOrigin-RevId: cc67f9e8cc54b0b8def4e9e4d8ee0dfb0e761040