Keep Jenkins docker args consistent with Stirling
Summary:
The way we expect the Stirling container (including PEM yamls) to be configured
has evolved. Since Jenkins is essentially running Stirling through a container (the dev_image),
we can--and should--treat it the same way.

Thus, use the same set of flags as we have in our scripts and yamls.

Test Plan: Jenkins

Reviewers: yzhao, #engineering, zasgar

Reviewed By: #engineering, zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D3293

GitOrigin-RevId: b5f46ff301bc5a2670b549d83d9f3dcc3553a43a