Add material-ui as a UI dependency, also update the npm offline packages.
Summary:
Since we are going to use material design for the Live view, switch to Material-UI for the base of our components instead. IMO There are several advanages for using Material-UI:
- Easy theming support
- Style and className encapsulation (no more verbose classnames like pixie-editor-content-text)
- Inline styles with JSS (defining the styles closer to markup)
- Large catalog of components

Our app might appear bloated now by using both bootstrap and Material-UI, but it shouldn't be hard to optimized via webpack code splitting and spliting up the bundles.

Test Plan: n/a

Reviewers: michelle, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3308

GitOrigin-RevId: 35a905070e100497bfb589fac03851a3a06957fd