PL-1381 Fix bug where most pods are labeled as "Terminated"
Summary:
Noticed an issue where pod statuses were labeled as terminated when they are actually running.
Turns out the issue is in Kelvin, when it runs RemoveDeadPods in state_manager.cc. Most of the pods sent to Kelvin don't actually belong to Kelvin, so all of those pods were marked as deleted when they shouldn't be.
Ideally, we want to remove this function. However, some more research is probably necessary: it looks like kube-proxy pods do not have a pod dir for some reason, although they _are_ being sent to the correct agents. Otherwise, it looks like all other pods are being redirected to the right place.

I also put ProcessPIDUpdates behind the "collects_data" flag, because Kelvin should not need to send PID updates to the metadata service. the PEM thats running on the same node as the Kelvin should already be doing that.

Test Plan: ran skaffold and checked that the terminated pods are actually terminated

Reviewers: oazizi, zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3326

GitOrigin-RevId: 58a74de516af029fc69ee2a0c083d4a5b5599ef5