Stirling: Switch to server-side tracing
Summary:
Based on learnings of how kubernetes networking is setup, it was decided that it's better to perform server-side tracing inside the cluster.

This will allow us to see how requests to a particular service is load-balanced onto the pods. In client-side tracing, we would be before the load-balancing point, and would not have access to this data.

lint

Test Plan: Updated existing tests, and ran stirling_wrapper manually on GKE.

Reviewers: yzhao, zasgar, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3323

GitOrigin-RevId: 699b396ef003a6fadd178b92cbd038aa4e93e90d