Fix agent queue bug: send on closed channel
Summary: we need some locking around the update agent queues in the metadata service. it is possible for the updateAgent goroutine to delete an agent, and close its channel, as we're trying to queue something on the agent's update channel.

Test Plan: ran skaffold, things seemed to run as normal.

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3336

GitOrigin-RevId: 9b248670471cc4dcdde19bd2438af01994403efe