Unify SourceNode logic for whether they have batches remaining
Summary:
From PL-1318, I am moving the logic for GRPC timeout from the source node to the exec graph.
When the graph times out waiting for work to do, it will trigger the stalled sources to send a
0-row end of stream RowBatch.
In order to make that easier, and remove the duplicated/slightly inconsistent logic across the source nodes,
I raised up the sent_eos_ logic to be set at the exec node level and HasBatchesRemaining will always be
dependent on whether EOS was sent or not. This is a different concept from NextBatchReady, which indicates
that the source is ready to produce the next batch.

Test Plan: existing

Reviewers: michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PL-1318

Differential Revision: https://phab.corp.pixielabs.ai/D3337

GitOrigin-RevId: 76635e5d99989223078e347ad29069a81f0d16a7