Down grade logs
Summary:
Use VLOG or LOG_EVERY_N

The rule to use VLOG is that the message is informational but does not signaling limitation of the
code.

LOG_EVERY_N is used when the code has limitations, for which we want to see the occurrence of the
situations that reveal them, but at a much lower frequency.

Test Plan: Jenkins

Reviewers: oazizi, zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3339

GitOrigin-RevId: ec4298e4c3533548c275b4046f43f739d5b98956