Add locking for cold batches in carnot table
Summary:
we were running into some concurrency issues where we were trying to read from the cold batches, because we were trying to delete from them at the same time.
there are a bunch of places where we're accessing the coldbatches that need to have a lock.

Test Plan: everything appears to run as normal, but i havent run long enough so that the data starts getting expired

Reviewers: zasgar, philkuz, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3342

GitOrigin-RevId: ecd63357a78e0ceb855618b7751d8cb7494b91d5