PL-1331: Drop _attr_* columns in MemorySink's
Summary:
The following query:
```
t1=px.DataFrame(table='process_stats', select=['upid'])
t1.pod_name = t1.ctx['pod_name']
px.display(t1)
```
would previously produce both `pod_name` and `_attr_pod_name` columns in the memorysink.

Now it produces only `upid` and `pod_name` as expected.

This fix simply drops any columns in the memory sink with the `_attr` prefix. However, I think our current handling of Metadata is a bit convoluted and overcomplicated and could probably use simplification.

Note that if the parent of the MemorySink is a map column, then this adds a drop after the map, and then later optimizations remove all of the colexprs in the Map, so we're left with two duplicate maps (one for the original map that was pruned down and one for the drop that was turned into a map). But I think this should be handled in a separate optimization where we remove no-ops/duplicate ops.

Test Plan: Added compiler level test, and tried it out in UI.

Reviewers: #engineering, nserrino

Reviewed By: #engineering, nserrino

Subscribers: philkuz, nserrino

Differential Revision: https://phab.corp.pixielabs.ai/D3325

GitOrigin-RevId: 1c2a5b53b67e4bb50c29f574dda5d5ac27194e53