SocketInfoDB: Cached access to SocketInfo.
Summary:
Turns out that the NetlinkProber can be quite expensive, even when going through the SocketProberManager.

This SocketInfoDB is a cached interface around SocketProberManager to reduce the cost of querying for connection info.

The integration into Stirling will be part of a future diff.

Test Plan: Tested using the DummyContainerTest.

Reviewers: zasgar, yzhao, #engineering

Reviewed By: yzhao, #engineering

JIRA Issues: PL-1388

Differential Revision: https://phab.corp.pixielabs.ai/D3355

GitOrigin-RevId: cb5180d51ff8561da203c4a3f5117d3506aa36b5