Jenkins: Cancel master build if a later build is in the queue
Summary:
Currently in jenkins-master, we build everything thats enqueued. however, if there are many pending builds, we might as well just cancel all of the pending builds in the middle and build the latest item in the queue.
unfortunately, theres no nice plugin that automatically cancels pending jenkins builds for us, although there are some open JIRA issues for this that havent been resolved since 2017.
what we can do is check at the beginning of each build whether another jenkins job exists in the queue. if there is, the build can just end early so that the later job can execute.

Test Plan: tried it out in a jenkins-cancel branch with a test jenkins job (https://jenkins.pixielabs.ai/job/pixielabs-jenkins-test/)

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Subscribers: oazizi

Differential Revision: https://phab.corp.pixielabs.ai/D3364

GitOrigin-RevId: 0b302bb1a0439dc73c5388c80131770b4b75d69b