Socket info tool
Summary:
This is a tool used for querying for sockets within a network namespace, using the netlink socket prober.

This was used to do some debug and comparison against other tools.

Useful to keep around for future debugging. Also a good reference guide of how to use the socket_info classes.

Test Plan: Ran it.

Reviewers: yzhao, zasgar, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3382

GitOrigin-RevId: 5beb51f778d9f090e420d5710d6b2b56983c16e0