PL-1341: QLObject Args should be QLObjects, not IRNodes
Summary:
This change is in support of adding px.flags to the QL.
For that case, px.flags takes an argument type of "type", which has a QLObject
in our system but not an IRnode. Currently, arguments to funcs/dataframes are
stored as IRNodes, which is not scalable to situations like this as well as other things like
funcs.

This change moves over ArgMap and ParsedArgs to use QLObjects instead of IRnodes.
It also adds some helpers/test methods to reduce boiler plate.
Most notably, it adds ParsedArgs::GetArgAs<> which is templatized to reduce the proliferation
of type casts and matches in arg parsing logic. The necessary methods on IRNode classes
can be added as-needed.

Test Plan: existing

Reviewers: philkuz, jamesbartlett, zasgar, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PL-1341

Differential Revision: https://phab.corp.pixielabs.ai/D3399

GitOrigin-RevId: 605f4bd75f99406d411f721dab4ae5cd3b966cb6