PL-1318: Make ExecGraph::Execute yieldable so that we don't hang waiting for source data
Summary:
This is the first diff for PL-1318 which makes the exec graph yieldable.
Currently we don't have any nodes that fit the yieldable pattern, however GRPC Source node
is going to be changed so that it no longer has a blocking timeout. In the next diff, the
GRPC router will call Continue() on the exec_graph whenever it enqueues a row batch.

However, TSAN is failing for the unit test that I added and I am not sure how to fix it. It seems like
we should have a test for this, but with TSAN sometimes the timeout on the condition variable occurs
before Continue() is able to be invoked by the test. I would appreciate any thoughts people have about
how to test this.

Test Plan: see above..

Reviewers: zasgar, michelle, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PL-1318

Differential Revision: https://phab.corp.pixielabs.ai/D3351

GitOrigin-RevId: af65913daef72f0a2c9fb7a0a9866e783d56930d