Fix Parser syntax error on funcdef with no args
Summary:
Previously, a query like:
```
def func():
    return 'a'

func
```
would have printed a SyntaxError to console but not propagated the error to our wrapper of the libpypa parser.
Now all SyntaxErrors in the parser get propagated to our wrapper, and it doesn't throw a SyntaxError on funcs with no args.

Test Plan: Add parser tests.

Reviewers: #engineering, philkuz, nserrino

Reviewed By: #engineering, philkuz

Differential Revision: https://phab.corp.pixielabs.ai/D3410

GitOrigin-RevId: ef15e0083fe2808176168caaa9d4786542ab393f