Optimized LittleEndianByteStrToInt
Summary:
Pass the number of bytes as a template parameter, so that the compiler can perform loop unrolling optimizations.

Honestly, this optimization is not expected to have any direct performance impact in our use cases, since the byte conversions are not a significant part of our processing. But it was an easy optimization.

Test Plan: Existing tests

Reviewers: yzhao, zasgar, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3424

GitOrigin-RevId: 797158ec288aea5a3d5959b068292bbb7a37c462