Cassandra: types extractor
Summary:
This is a library of tools to extract native types out of cassandra frames.

Currently standalone, but will be used by the cassandra stitcher, when it tries to interpret the contents of frames.

Test Plan: An extensive set of tests have been added, although I have left a few cases as future work because this diff is already very large.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

JIRA Issues: PL-711

Differential Revision: https://phab.corp.pixielabs.ai/D3446

GitOrigin-RevId: baa2210b5c1bf5c682e34d020ed357f1ff2b4b31