Repr/BytesToString: use template parameters
Summary:
In the process of the Cassandra work, I realized I needed a way to represent binary data, which would be displayed in the table output.

We already had Repr(), but I wanted to play with the formatting.

Ended up restructuring Repr() a bit, and renamed it to BytesToString() to make it clearer what it is doing.

Also switched it to use template arguments, for run-time efficiency, since we always know which format we want to use at compile-time.

Test Plan: Tests in utils

Reviewers: yzhao, zasgar, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D3475

GitOrigin-RevId: 6185d0aba68eeb825514c3e8dd5667de932fc1c1