Cassandra: Handle QUERY opcode
Summary:
Process the query frame to extract out all the fields.

For now, just output the query string. Future work to incorporate some of the other parsed fields into the output.

Test Plan: Use a captured QUERY frame and see if we can parse it. There are so many flags and variations, that we probably need to create a fuzzer at some point.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

JIRA Issues: PL-711

Differential Revision: https://phab.corp.pixielabs.ai/D3466

GitOrigin-RevId: 3fc66e08061754e462f9b7273552d8ca180043a8