PL-754: Adding support for Rolling windowing in compiler and query language.
Summary:
Adds a .rolling function to support windowed aggregates, eg.
```
t1 = px.DataFrame(..., select=['time_', 'col1'])
t1 = t1.rolling('2s').agg(...)
```
The rolling function also supports an `on` parameter to specify which column to window on, however it currently only supports windowing on `time_`. `t1.rolling('2s')` is equivalent to `t1.rolling('2s', on='time_')` and `t1.rolling(10, on='col1')` is currently not supported but will be in the future.

Currently, the proto spec is undefined so attempting to compile to a proto will result in an Unimplemented error.

This PR also adds a GroupAcceptorIR that serves as an interface for any IR that can accept groups from a GroupBy op. This PR changes BlockingAgg to subclass GroupAcceptorIR and also the new RollingIR is also a subclass of GroupAcceptorIR. This allows for both `t1.groupby(...).rolling(...).agg(...)` and `t1.groupby(...).agg(...)` to be handled by the Merge rule.

Additionally, this PR adds support for the RollingIR window size parameter to accept time strings as above, or compile time expr evaluation such as `t1.rolling(1 + px.now())`.

Currently, the RollingIR is left in the graph, but once the spec for the proto is known it will likely need to be merged into the Agg.

Test Plan:
- Add tests to ensure new `MergeGroupByIntoGroupAcceptorRule` works for both `RollingIR` and `BlockingAggIR`.
- Add tests to ensure transition from `ConvertMemSourceStringTimesRule` to `ConvertStringTimeRule` still works for mem source and additionally works for Rolling now.
- Add tests that RollingIR node gets created properly.
- Add tests to check that compile time expr eval works for new Rolling op.

Reviewers: #engineering, philkuz, nserrino

Reviewed By: #engineering, nserrino

Subscribers: nserrino, philkuz

JIRA Issues: PL-754

Differential Revision: https://phab.corp.pixielabs.ai/D3365

GitOrigin-RevId: c59cfb51bb9f11d444ee6551c6092ed17710aa22